Galería
Se ha modificado como se ha añadido la galería a Ad.js.  Se ha utilizado el componente Gallery.js.
Se ha eliminado Advertisement por repetición.
