+++ src/App.js
@@ -9,8 +9,6 @@ import Registration from "screens/UserFlow/Registration/Registration";
 import RecoverPassword from "screens/UserFlow/RecoverPassword/RecoverPassword";
 import Profile from "screens/UserFlow/Profile/Profile";
 
-import Advertisement from "screens/Advertisement/Advertisement";
-
 const App = () => {
 	return (
 		<>
@@ -21,8 +19,6 @@ const App = () => {
 				<Route exact path="/register" component={Registration} />
 				<Route exact path="/profile" component={Profile} />
 				<Route exact path="/recover-password/:hash" component={RecoverPassword} />
-
-				<Route exact path="/ads" component={Advertisement} />
 			</Switch>
 		</>
 	);
